+++ src/minter.js
@@ -4,7 +4,7 @@ import PostTx from './api/post-tx';
 import EstimateCoinSell from './api/estimate-coin-sell';
 import EstimateCoinBuy from './api/estimate-coin-buy';
 import EstimateTxCommission from './api/estimate-tx-commission';
-import issueCheck from './issue-check';
+import issueCheck, {decodeCheck} from './check';
 
 /**
  * @param {Object} [options]
@@ -27,4 +27,6 @@ export default function (options) {
     this.estimateTxCommission = new EstimateTxCommission(apiInstance);
 
     this.issueCheck = issueCheck;
+
+    this.decodeCheck = decodeCheck;
 }