+++ test/check.test.js
@@ -1,20 +1,22 @@
 import {Buffer} from 'safe-buffer';
-import {issueCheck} from '~/src';
+import {issueCheck, decodeCheck} from '~/src';
 
+
+const VALID_CHECK = 'Mcf8a03101830f423f8a4d4e5400000000000000888ac7230489e80000b84149eba2361855724bbd3d20eb97a54ea15ad7dc28c1111b8dcf3bb15db26f874f095803cad9f8fc88b2b4eec9ba706325a7929be31b6ccfef01260791a844cb55011ba06c63ad17bfe07b82be8a0144fd4daf8b4144281fdf88f313205ceacf37fd877fa03c243ad79cab6205f4b753bd402c4cfa5d570888659090b2f923071ac52bdf75';
+const checkParams = {
+    privateKey: '2919c43d5c712cae66f869a524d9523999998d51157dc40ac4d8d80a7602ce02',
+    passPhrase: 'pass',
+    nonce: '1',
+    chainId: 1,
+    coin: 'MNT',
+    value: '10',
+    dueBlock: 999999,
+};
 
 describe('issueCheck()', () => {
-    const checkParams = {
-        privateKey: '2919c43d5c712cae66f869a524d9523999998d51157dc40ac4d8d80a7602ce02',
-        passPhrase: 'pass',
-        nonce: '1',
-        chainId: 1,
-        coinSymbol: 'MNT',
-        value: 10,
-        dueBlock: 999999,
-    };
     test('should work', () => {
         const check = issueCheck(checkParams);
-        expect(check).toEqual('Mcf8a03101830f423f8a4d4e5400000000000000888ac7230489e80000b84149eba2361855724bbd3d20eb97a54ea15ad7dc28c1111b8dcf3bb15db26f874f095803cad9f8fc88b2b4eec9ba706325a7929be31b6ccfef01260791a844cb55011ba06c63ad17bfe07b82be8a0144fd4daf8b4144281fdf88f313205ceacf37fd877fa03c243ad79cab6205f4b753bd402c4cfa5d570888659090b2f923071ac52bdf75');
+        expect(check).toEqual(VALID_CHECK);
     });
 
     test('should accept buffer private key', () => {
@@ -22,7 +24,7 @@ describe('issueCheck()', () => {
             ...checkParams,
             privateKey: Buffer.from(checkParams.privateKey, 'hex'),
         });
-        expect(check).toEqual('Mcf8a03101830f423f8a4d4e5400000000000000888ac7230489e80000b84149eba2361855724bbd3d20eb97a54ea15ad7dc28c1111b8dcf3bb15db26f874f095803cad9f8fc88b2b4eec9ba706325a7929be31b6ccfef01260791a844cb55011ba06c63ad17bfe07b82be8a0144fd4daf8b4144281fdf88f313205ceacf37fd877fa03c243ad79cab6205f4b753bd402c4cfa5d570888659090b2f923071ac52bdf75');
+        expect(check).toEqual(VALID_CHECK);
     });
 
     test('default dueBlock: 999999', () => {
@@ -68,4 +70,12 @@ describe('issueCheck()', () => {
             value: '123asd',
         })).toThrow();
     });
+});
+
+describe('decodeCheck()', () => {
+    const checkParamsWithoutSensitiveData = Object.assign({}, checkParams);
+    delete checkParamsWithoutSensitiveData.passPhrase;
+    delete checkParamsWithoutSensitiveData.privateKey;
+
+    expect(decodeCheck(VALID_CHECK)).toEqual(checkParamsWithoutSensitiveData);
 });